server/requirements.txt
flask
flask-sqlalchemy
flask-login
paho-mqtt
jwt
python-dotenv